Fix: Preserve milliseconds in time parsing for accurate pace calculations
Problem: Time format "00:05:43.420" was being parsed incorrectly - milliseconds
were truncated before pace calculation, causing inaccurate average pace values.

Changes:
- PHP: Changed intval() to floatval() for seconds component in parse_time_to_seconds()
- JavaScript: Changed parseInt() to parseFloat() for seconds in parseTimeToSeconds()
- Python: Changed int() to float() for seconds in parse_time_to_seconds()

This ensures millisecond precision is maintained throughout pace calculations.

+++ assets/js/chronotrack-live.js
@@ -1687,7 +1687,8 @@
             if (!timeStr || timeStr === '-') return 999999;
             const parts = timeStr.split(':');
             if (parts.length === 3) {
-                return parseInt(parts[0]) * 3600 + parseInt(parts[1]) * 60 + parseInt(parts[2]);
+                // Use parseFloat for seconds to preserve milliseconds (e.g., "43.420")
+                return parseInt(parts[0]) * 3600 + parseInt(parts[1]) * 60 + parseFloat(parts[2]);
             }
             return 999999;
         },
